cast added to avoid compiler warnings
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk@1924 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ trdp/src/common/trdp_utils.c
@@ -900,7 +900,7 @@ TRDP_ERR_T  trdp_requestSocket (
                         }
                         continue;   /* No, socket cannot join more MC groups */
                     }
-                    vos_printLog(VOS_LOG_INFO, "socket %d joined %s!\n", iface[lIndex].sock, vos_ipDotted(mcGroup));
+                    vos_printLog(VOS_LOG_INFO, "socket %d joined %s!\n", (int) iface[lIndex].sock, vos_ipDotted(mcGroup));
                 }
             }
